Improvements
Toggling of list menu is handled by Knockout.js, error handling method
for Google Maps was added, error methods for Wikipedia API and
FourSquare API were updated to jQuery documentation specifications. A
close menu button was added for mobile screens and specific css was
added to handle not displaying links when the API data is undefined —
the property requested does not exist.

The button had to be added because Knockout.js interferes with the DOM
and when I used the map div to register the clicks to close the menu
list the links in the Google Maps infowindow did not function when
clicked.